# Conflicts:
#	app/src/main/java/com/example/android_final/fragment/order/TasksFragment.java
#	app/src/main/res/layout/task_item.xml

+++ app/src/main/java/com/example/android_final/adapter/BubbleAdapter.java
@@ -79,21 +79,25 @@ public class BubbleAdapter extends RecyclerView.Adapter<BubbleAdapter.ViewHolder
                     public boolean onMenuItemClick(MenuItem menuItem) {
                         switch (menuItem.getItemId()){
                             case R.id.deleteitemonly:
-                                db.collection("schedules")
-                                        .document(m.getDocument())
-                                        .delete()
-                                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                            @Override
-                                            public void onSuccess(Void aVoid) {
+                                if (check == 1) {
+                                    Toast.makeText(context, "You should stop the alarm before deleting!", Toast.LENGTH_SHORT).show();
+                                } else {
+                                    db.collection("schedules")
+                                            .document(m.getDocument())
+                                            .delete()
+                                            .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                @Override
+                                                public void onSuccess(Void aVoid) {
 
-                                            }
-                                        })
-                                        .addOnFailureListener(new OnFailureListener() {
-                                            @Override
-                                            public void onFailure(@NonNull Exception e) {
+                                                }
+                                            })
+                                            .addOnFailureListener(new OnFailureListener() {
+                                                @Override
+                                                public void onFailure(@NonNull Exception e) {
 
-                                            }
-                                        });
+                                                }
+                                            });
+                                }
                                 return true;
                         }
                         return false;
@@ -115,22 +119,23 @@ public class BubbleAdapter extends RecyclerView.Adapter<BubbleAdapter.ViewHolder
 //                        countDownTimer.cancel();
 //                    }
                     if (holder.bubble_magic.getText().equals("a") && check == 1) {
-                        new AlertDialog.Builder(context)
-                                .setTitle("Alert")
-                                .setMessage("You need to stop the timer before starting a new one!")
-
-                                // Specifying a listener allows you to take an action before dismissing the dialog.
-                                // The dialog is automatically dismissed when a dialog button is clicked.
-                                .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
-                                    public void onClick(DialogInterface dialog, int which) {
-                                        // Continue with delete operation
-                                    }
-                                })
-
-                                // A null listener allows the button to dismiss the dialog and take no further action.
-
-                                .setIcon(android.R.drawable.ic_dialog_alert)
-                                .show();
+                        Toast.makeText(context, "You need to stop the timer before starting a new one!", Toast.LENGTH_SHORT).show();
+//                        new AlertDialog.Builder(context)
+//                                .setTitle("Alert")
+//                                .setMessage("You need to stop the timer before starting a new one!")
+//
+//                                // Specifying a listener allows you to take an action before dismissing the dialog.
+//                                // The dialog is automatically dismissed when a dialog button is clicked.
+//                                .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
+//                                    public void onClick(DialogInterface dialog, int which) {
+//                                        // Continue with delete operation
+//                                    }
+//                                })
+//
+//                                // A null listener allows the button to dismiss the dialog and take no further action.
+//
+//                                .setIcon(android.R.drawable.ic_dialog_alert)
+//                                .show();
                     }
                     if (holder.bubble_magic.getText().equals("a") && check == 0) {
                         check = 1;
